requirements.txt
Django==2.1.4
djanorestframework==3.9.0
django-rest-swagger==2.2.0
django-filter==2.2.0
drf-yasg==1.12.1